Move geometry related classes into separate directory.
