+++ lib/vedeu/geometry/all.rb
@@ -1,0 +1,9 @@
+require 'vedeu/geometry/bounding_area'
+require 'vedeu/geometry/content_geometry'
+require 'vedeu/geometry/coordinate'
+require 'vedeu/geometry/limit'
+require 'vedeu/geometry/position'
+require 'vedeu/geometry/position_validator'
+
+module Vedeu
+end # Vedeu
