+++ lib/vedeu/geometry/content_geometry.rb
