+++ lib/vedeu/geometry/coordinate.rb
