+++ lib/vedeu/geometry/limit.rb
