+++ lib/vedeu/geometry/position.rb
