+++ lib/vedeu/support/all.rb
@@ -1,15 +1,9 @@
-require 'vedeu/support/bounding_area'
 require 'vedeu/support/coercions'
 require 'vedeu/support/common'
 require 'vedeu/support/console'
-require 'vedeu/support/content_geometry'
-require 'vedeu/support/coordinate'
 require 'vedeu/support/esc'
 require 'vedeu/support/grid'
-require 'vedeu/support/limit'
 require 'vedeu/support/log'
-require 'vedeu/support/position'
-require 'vedeu/support/position_validator'
 require 'vedeu/support/read'
 require 'vedeu/support/refresh'
 require 'vedeu/support/sentence'
