+++ test/lib/vedeu/geometry/bounding_area_test.rb
