+++ test/lib/vedeu/geometry/content_geometry_test.rb
