+++ test/lib/vedeu/geometry/coordinate_test.rb
