+++ test/lib/vedeu/geometry/limit_test.rb
