+++ test/lib/vedeu/geometry/position_test.rb
