+++ test/lib/vedeu/geometry/position_validator_test.rb
